Refactor #342 - cleanup startup.init package
