+++ app/src/debug/kotlin/com/nlab/reminder/apps/startup/init/FlipperInitializer.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.nlab.reminder.internal.common.android.startup.init
+package com.nlab.reminder.apps.startup.init
 
 import android.content.Context
 import androidx.startup.Initializer
@@ -27,7 +27,7 @@ import com.facebook.flipper.plugins.inspector.InspectorFlipperPlugin
 import com.facebook.flipper.plugins.leakcanary2.FlipperLeakEventListener
 import com.facebook.flipper.plugins.leakcanary2.LeakCanary2FlipperPlugin
 import com.facebook.soloader.SoLoader
-import com.nlab.reminder.internal.common.android.startup.EmptyDependencies
+import com.nlab.reminder.apps.startup.EmptyDependencies
 import leakcanary.LeakCanary
 
 /**
